requirements.txt
streamlit==0.89.0
streamlit-folium==0.4.0
streamlit-metrics==0.1.0
plotly==5.2.1
plotly-express==0.4.1
pandas==1.2.4
numpy==1.20.1
matplotlib==3.3.4
folium==0.12.1
DateTime==4.3
seaborn==0.11.1
Pillow==8.2.0
path==15.1.2